+++ changelog.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented here.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `createUser()` to `UserService`
 
 ## [1.2.0]
 ### Added